address sanitiyer

+++ btrblocks/scheme/double/Alp.cpp
@@ -533,8 +533,8 @@ void Alp::decompress(DOUBLE* dest,
   if (col_struct.exceptions_count > 0) {
     thread_local vector<vector<DOUBLE>> exceptions_v;
     thread_local vector<vector<INTEGER>> patches_v;
-    auto exceptions_ptr = get_level_data(exceptions_v, col_struct.exceptions_count + SIMD_EXTRA_ELEMENTS(INTEGER), level);
-    auto patches_ptr = get_level_data(patches_v, col_struct.exceptions_count + SIMD_EXTRA_ELEMENTS(DOUBLE), level);
+    auto exceptions_ptr = get_level_data(exceptions_v, col_struct.exceptions_count + SIMD_EXTRA_ELEMENTS(DOUBLE), level);
+    auto patches_ptr = get_level_data(patches_v, col_struct.exceptions_count + SIMD_EXTRA_ELEMENTS(INTEGER), level);
 
     IntegerScheme& exception_positions_scheme =
         IntegerSchemePicker::MyTypeWrapper::getScheme(col_struct.exceptions_positions_scheme);
